Set up file for common Vue functions
