+++ Frontend/src/components/admin/CoopReviewPageNewCoopItem.vue
@@ -50,39 +50,12 @@ export default {
   created: function() {
     var bytes = this.coop.studentReports[0].data;
 
-    let blob = this.b64toBlob(bytes, "application/pdf");
+    let blob = this.$common.b64toBlob(bytes, "application/pdf");
     this.offerLetterURL = window.URL.createObjectURL(blob);
   },
   methods: {
     openPDF: function() {
       window.open(this.offerLetterURL);
-    },
-    // found on SO: https://stackoverflow.com/a/57767153
-    b64toBlob: function(content, contentType) {
-      contentType = contentType || "";
-      const sliceSize = 512;
-      // method which converts base64 to binary
-      const byteCharacters = window.atob(content);
-
-      const byteArrays = [];
-      for (
-        let offset = 0;
-        offset < byteCharacters.length;
-        offset += sliceSize
-      ) {
-        const slice = byteCharacters.slice(offset, offset + sliceSize);
-        const byteNumbers = new Array(slice.length);
-        for (let i = 0; i < slice.length; i++) {
-          byteNumbers[i] = slice.charCodeAt(i);
-        }
-        const byteArray = new Uint8Array(byteNumbers);
-        byteArrays.push(byteArray);
-      }
-
-      // statement which creates the blob
-      return new Blob(byteArrays, {
-        type: contentType
-      });
     }
   }
 };
